+++ c/vm.h
@@ -1,6 +1,8 @@
 #ifndef VM_H_
 #define VM_H_
 
+#include <stdint.h>
+
 #define MAX_MEM 1024
 
 typedef enum {
@@ -25,11 +27,11 @@ typedef struct {
     error error;
     int pc;
     int mem_size;
-    int mem[MAX_MEM];
+    int64_t mem[MAX_MEM];
 } vm;
 
 vm new_vm(void);
-vm make_vm(int mem[], int mem_size);
+vm make_vm(int64_t mem[], int mem_size);
 void print_vm(const vm* vm);
 void step(vm* vm);
 void run(vm* vm);
